requirements.txt
numpy
pandas
matplotlib
seaborn
tensorflow
joblib
pyyaml